requirements.txt
torch==1.4.0
onnx==1.5.0
torchvision==0.5.0
numpy==1.17.2
pretrainedmodels==0.7.4
opencv_python==4.1.1.26
Pillow==7.0.0
